Split out iterators and `WithTargetFn` into their own files
This moves all the iterators to `iter.rs` and `WithTargetFn` into it's own module file.
